refactor: Apply context to orderSort #50
