+++ components/select/OrderSort.tsx
@@ -1,19 +1,19 @@
 import styles from 'styles/Pulldown.module.css'
 import {memo, useContext} from "react"
-import { TodoContext } from 'providers/TodoProvider';
+import { SortContext, TodoContext } from 'providers/TodoProvider';
 
 
-type Props = {
-  orderSort: string
-  onChangeOrderSort: (e: React.ChangeEvent<HTMLSelectElement>) => void
-}
-
-export const OrderSortButton:React.FC<Props> = memo((props) => {
+export const OrderSortButton:React.VFC = memo(() => {
   OrderSortButton.displayName = 'OrderSortButton';
   console.log('Render Order Sort')
 
-  const {orderSort, onChangeOrderSort } = props;
   const { isDisabled } = useContext(TodoContext)
+  const {orderSort, setOrderSort} = useContext(SortContext)
+
+  const onChangeOrderSort = (e: React.ChangeEvent<HTMLSelectElement>) => {
+    e.preventDefault()
+    setOrderSort(e.target.value)
+  }
 
   return (
     <>
